requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==19.3.0
Automat==0.8.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
click==8.0.1
cloud-init==20.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
loader==2017.9.11
Logbook==1.5.3
logcc==2018.3.1
MarkupSafe==2.0.1
more-itertools==4.2.0
multidict==5.1.0
netifaces==0.10.4
numpy==1.21.1
oauthlib==3.1.0
pandas==1.3.1
pendulum==2.1.2
pexpect==4.6.0
pony==0.7.14
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.4.2
pymacaroons==0.13.0
pymongo==3.11.4
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.19.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==0.24.2
scipy==1.7.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0
sos==4.0
soupsieve==2.2.1
ssh-import-id==5.10
systemd-python==234
termcc==2019.8.2
terminaltables==3.1.0
threadpoolctl==2.2.0
Twisted==18.9.0
typing-extensions==3.10.0.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.4.3
wadllib==1.3.3
websockets==9.1
Werkzeug==2.0.1
wrap==2018.3.1
yarl==1.6.3
zipp==1.0.0
zope.interface==4.7.1
